fix/spendings-from
