+++ src/public/feat1-1.php
@@ -1,0 +1,6 @@
+<?php
+require_once __DIR__ . '/../vendor/autoload.php';
+use App\SpendingCalculator;
+
+$spendings = new SpendingCalculator("root", "password");
+var_dump($spendings->fetchAllspendings());
